Allow the 'created' attribute to be accessed for favorited items
This allows you to check the date an item was added, through a datetime
object exposed as a user_date_added attribute. The 'created' key gets
moved into the 'item' array in the for simplicity. This also changes
the add_remove favorite tests to verify that the date is correct.