Support logging in with OAuth
All the known api tokens for the old login end point
have expired, so now you log in by logging in through
a website, and entering a code specific to your session.

This commit also updates the README to show the correct
documentation for the 0.7.x branch, and includes some
extra documentation on how to log in with OAuth.