Update favorites.tracks to accept order and orderDirection params
I've defaulted these to what appears to be the default for the API, but
adding them allows users to pass in e.g. order=DATE orderDirection=DESC
to get the tracks in the date they were added to favorites. There are
likely other possibilities for the order but I'm not sure what they are
(;